+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 FIXES:
 [#142](https://github.com/sleuth-io/terraform-provider-sleuth/pull/142) Fix provider value case from API
 
+EHANCEMENTS:
+[#143](https://github.com/sleuth-io/terraform-provider-sleuth/pull/143) Add Shortcut (ex Clubhouse) as Incident Impact Source
+
 ## 0.4.5 (September 26, 2023)
 EHANCEMENTS:
 - [#140](https://github.com/sleuth-io/terraform-provider-sleuth/pull/140) Update OpsGenie Incident Impact Source docs