+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.4.6 (unreleased)
+## 0.4.6 (October 23, 2023)
 FIXES:
 [#142](https://github.com/sleuth-io/terraform-provider-sleuth/pull/142) Fix provider value case from API